listening-comp/backend/requirements.txt
chromadb
streamlit
boto3
youtube_transcript_api
aiofiles>=23.2.1
sentence-transformers>=2.5.1
langdetect>=1.0.9
googletrans>=3.0.0